Fix staticcheck issues

+++ initialization/initialization_test.go
@@ -69,7 +69,7 @@ func TestInitialize(t *testing.T) {
 	m, err := LoadMetadata(opts.DataDir)
 	r.NoError(err)
 	r.NotNil(m.Nonce)
-	r.NoError(verifying.VerifyPow(*m.Nonce, uint64(opts.NumUnits), uint64(cfg.BitsPerLabel), nodeId, commitmentAtxId))
+	r.NoError(verifying.VerifyPow(*m.Nonce, opts.NumUnits, cfg.BitsPerLabel, nodeId, commitmentAtxId))
 }
 
 func TestInitialize_PowOutOfRange(t *testing.T) {
@@ -103,7 +103,7 @@ func TestInitialize_PowOutOfRange(t *testing.T) {
 	m, err := LoadMetadata(opts.DataDir)
 	r.NoError(err)
 	r.NotNil(m.Nonce)
-	r.NoError(verifying.VerifyPow(*m.Nonce, uint64(opts.NumUnits), uint64(cfg.BitsPerLabel), nodeId, commitmentAtxId))
+	r.NoError(verifying.VerifyPow(*m.Nonce, opts.NumUnits, cfg.BitsPerLabel, nodeId, commitmentAtxId))
 
 	// check that the found nonce is outside of the range for calculating labels
 	r.Less(uint64(cfg.MinNumUnits)*cfg.LabelsPerUnit, *m.Nonce)
@@ -136,7 +136,7 @@ func TestInitialize_ContinueWithLastPos(t *testing.T) {
 	m, err := LoadMetadata(opts.DataDir)
 	r.NoError(err)
 	r.NotNil(m.Nonce)
-	r.NoError(verifying.VerifyPow(*m.Nonce, uint64(opts.NumUnits), uint64(cfg.BitsPerLabel), nodeId, commitmentAtxId))
+	r.NoError(verifying.VerifyPow(*m.Nonce, opts.NumUnits, cfg.BitsPerLabel, nodeId, commitmentAtxId))
 
 	// trying again returns same nonce
 	origNonce := *m.Nonce
@@ -201,7 +201,7 @@ func TestInitialize_ContinueWithLastPos(t *testing.T) {
 	r.LessOrEqual(uint64(cfg.MinNumUnits)*cfg.LabelsPerUnit, *m.LastPosition)
 
 	r.LessOrEqual(uint64(cfg.MinNumUnits)*cfg.LabelsPerUnit, *m.Nonce)
-	r.NoError(verifying.VerifyPow(*m.Nonce, uint64(opts.NumUnits), uint64(cfg.BitsPerLabel), nodeId, commitmentAtxId))
+	r.NoError(verifying.VerifyPow(*m.Nonce, opts.NumUnits, cfg.BitsPerLabel, nodeId, commitmentAtxId))
 
 	// lastPos sets lower bound for searching for nonce if none was found
 	lastPos := *m.Nonce + 10
@@ -228,7 +228,7 @@ func TestInitialize_ContinueWithLastPos(t *testing.T) {
 	r.LessOrEqual(lastPos, *m.LastPosition)
 
 	r.Less(lastPos, *m.Nonce)
-	r.NoError(verifying.VerifyPow(*m.Nonce, uint64(opts.NumUnits), uint64(cfg.BitsPerLabel), nodeId, commitmentAtxId))
+	r.NoError(verifying.VerifyPow(*m.Nonce, opts.NumUnits, cfg.BitsPerLabel, nodeId, commitmentAtxId))
 }
 
 func TestReset_WhileInitializing(t *testing.T) {
@@ -520,7 +520,7 @@ func TestInitialize_MultipleFiles(t *testing.T) {
 	m, err := LoadMetadata(opts.DataDir)
 	r.NoError(err)
 	r.NotNil(m.Nonce)
-	r.NoError(verifying.VerifyPow(*m.Nonce, uint64(opts.NumUnits), uint64(cfg.BitsPerLabel), nodeId, commitmentAtxId))
+	r.NoError(verifying.VerifyPow(*m.Nonce, opts.NumUnits, cfg.BitsPerLabel, nodeId, commitmentAtxId))
 	oneFileNonce := *m.Nonce
 
 	for numFiles := uint32(2); numFiles <= 16; numFiles <<= 1 {
