+++ verifying/verifying_test.go
@@ -76,7 +76,7 @@ func TestVerifyPow(t *testing.T) {
 	m, err := initialization.LoadMetadata(opts.DataDir)
 	r.NoError(err)
 
-	r.NoError(VerifyPow(*m.Nonce, uint64(opts.NumUnits), uint64(cfg.BitsPerLabel), nodeId, commitmentAtxId))
+	r.NoError(VerifyPow(*m.Nonce, opts.NumUnits, cfg.BitsPerLabel, nodeId, commitmentAtxId))
 }
 
 // TestLabelsCorrectness tests, for variation of label sizes, the correctness of
